slight change in formatting of last commit
git-svn-id: svn://clusterresources.com/torque/branches/2.3-fixes@2934 8f940c70-5916-0410-ac46-d1fa2fa6ea29

+++ src/server/node_manager.c
@@ -2722,7 +2722,7 @@ static int number(
   int i = 0;
   char *str = *ptr;
 
-  while (isdigit(*str) && (size_t)(i+1) < sizeof holder)
+  while (isdigit(*str) && (unsigned int)(i + 1) < sizeof holder)
     holder[i++] = *str++;
 
   if (i == 0)
